// src/lib/api/types/api.ts
export interface components {
  schemas: {
// ...
    ResourceType:
      | "user"
      | "role"
      | "customer"
      | "supplier"
      | "workshop"
      | "worker"
      | "garment-type"
      | "garment-type-image"
      | "material-type"
      | "material"
      | "unit"
      | "action-type"
      | "action"
      | "size"
      | "garment-template"
      | "quotation"
      | "quotation-item"
      | "purchase-order"
      | "specification"
      | "production-assignment"
      | "progress-history"
      | "payment"
      | "account-receivable"
      | "account-payable"
      | "account-type"
      | "chart-account"
      | "movement"
      | "audit-log"
      | "measurement-control";
    /**
     * @description Nombre de la acción
     * @enum {string}
     */
    ActionType:
      | "read"
      | "create"
      | "update"
      | "activate"
      | "deactivate"
      | "export"
      | "*";
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}